requirements.txt
tifffile
streamlit-image-zoom
scikit-image
matplotlib
openpyxl
zarr
oiffile